Folderless Group Hack:
With a lot of effort, I seem to finally have added a folderless group by hacking the pbxproj file. I'm not sure how stable the project is after that as I just copied a GUID from another pbxproj file. It could be an issue in the long run, but I don't have much of an option here as it seems Apple wants to deprecate this feature and it seems to be impossible to create a folderless group in XCode 16.4 via xcode even though the release notes state that it should be possible to add one by ctrl-clicking in the project file explorer and then pressing the option key. I tried all possible variants of this but none works.

I'm really tired with Apple removing good features from XCode, like the two-column diff view they removed before (reimplemented, but not as useful as before). I hope Apple get's enough criticism about this so they'll eventually fix back the folderless groups support as I don't really want an actual "scripts" folder and I don't want the scripts to be mixed with the other files in the root level dir in the xcode project file list. What I want is to have the scripts being in the root folder but I don't want to see this cluttered list of files in XCode, but I want to have it well organized in xcode instead. A neat logical organization of files in XCode and the practicality of having scripts and such in the root folder which makes using them easier than having to write "scripts/fetch-dependencies.py" every time you want to run it.